requirements.txt
asgiref==3.7.2
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-environ==0.11.2
idna==3.7
mysqlclient==2.2.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
